+++ lib/resp_parser.h
@@ -16,6 +16,6 @@ int resp_process_packet(struct data_wrap *dw, struct resp_response*);
 void resp_print(struct resp_response *r);
 const char *resp_error(struct resp_response *r);
 struct resp_response *resp_alloc(void);
-char *resp_bulkstring_array_fetch(struct resp_response *r, int idx);
+char *resp_bulkstring_array_fetch(struct resp_response *r, unsigned long idx);
 long resp_array_length(struct resp_response *r);
 enum RESP_TYPE resp_type(struct resp_response *r);
